Added reference implementation
WinFyre is a very simple implementation of FyreVM through the Zifmia
wrapper. It has a sample game file embedded in the project that shows
how to use the EngineWrapper class and write text to different channels.
The output is shown in JSON, but can also be read from a
Dictionary<string,string> Name-Value-Pair list as well as from XML
(commented out by default in the EngineWrapper code).